json update

+++ app.py
@@ -1,5 +1,5 @@
 import urllib.request, urllib
-import zipfile, csv, redis, cherrypy, os, ast
+import zipfile, csv, redis, cherrypy, os, ast, json
 from jinja2 import Environment, FileSystemLoader
 
 
@@ -8,6 +8,9 @@ env=Environment(loader=FileSystemLoader(CUR_DIR), trim_blocks=True)
 
 
 class Index(object):
+    # Connect to redis database
+    r_db = redis.StrictRedis('localhost', 6379, charset="utf-8", decode_responses=True)
+    
     @cherrypy.expose()
     def index(self):
         template = env.get_template('index.html')
@@ -19,47 +22,48 @@ class Index(object):
     def download(self, day,month,year):
         template = env.get_template('table.html')
 
-        day,month,year = [int(day), int(month), int(year)]
+        date = '%s%s%s' % (day, month, year)
 
-        url = 'http://www.bseindia.com/download/BhavCopy/Equity/EQ%d%d%d_CSV.ZIP' % (day, month, year)
-        r_db = redis.StrictRedis('localhost', 6379, charset="utf-8", decode_responses=True)
-        r_db.flushall()
+        #condition to check if bhavData already exist in database for given date
+        if not self.r_db.hexists('bhavData', date):
+            url = 'http://www.bseindia.com/download/BhavCopy/Equity/EQ%s_CSV.ZIP' % (date)
 
-        try:
-            tmp_file = urllib.request.urlretrieve(url)
-        except urllib.error.HTTPError:
-            return template.render(data='')
-        archive = zipfile.ZipFile(tmp_file[0])
-        file = archive.extract(archive.namelist()[0])
+            try:
+                tmp_file = urllib.request.urlretrieve(url)
+            except urllib.error.HTTPError:
+                return template.render(data='')
+            archive = zipfile.ZipFile(tmp_file[0])
+            file = archive.extract(archive.namelist()[0])
 
-        with open(file, encoding='utf-8') as data:
-            csvfile = csv.reader(data)
-            next(csvfile)
-            for row in csvfile:
-                c_name = row[1].strip()
-                c_data = {'name': c_name, 'code': int(row[0]), 'open': float(row[4]), 'high': float(row[5]),
-                          'low': float(row[6]), 'close': float(row[7])}
-                r_db.rpush('list', c_data)
+            with open(file, encoding='utf-8') as data:
+                csvfile = csv.reader(data)
+                next(csvfile)
+                li = []
+                for row in csvfile:
+                    c_name = row[1].strip()
+                    c_data = {'name': c_name, 'code': int(row[0]), 'open': float(row[4]), 'high': float(row[5]),
+                              'low': float(row[6]), 'close': float(row[7])}
+                    li.append(c_data)
+                data = json.dumps(li)
+                self.r_db.hset('bhavData', date, data)
 
+        #get the top 10 companies and render
+        data = self.r_db.hget('bhavData', date)
+        bhav_data = ast.literal_eval(data)
+        return template.render(data=bhav_data[:10], date=date)
 
-        top_10 = r_db.lrange('list',0,9)
-        top = [ast.literal_eval(x) for x in top_10]
-        return template.render(data=top)
 
     @cherrypy.expose()
-    def search(self, name):
+    def search(self, name, date):
         template = env.get_template('table.html')
-        r_db = redis.StrictRedis('localhost', 6379, charset="utf-8", decode_responses=True)
+        comp = []
+        data = self.r_db.hget('bhavData', date)
+        bhav_data = ast.literal_eval(data)
+        for company in bhav_data:
+            if name.upper() in company['name']:
+                comp.append(company)
 
-        data=[]
-        c_name = name.upper()
-        for i in range(r_db.llen('list')):
-            dic = r_db.lindex('list',i)
-            c_data = ast.literal_eval(dic)
-            if c_name in c_data['name']:
-                data.append(c_data)
-
-        return template.render(data=data)
+        return template.render(data=comp)
 
 
 if __name__ == '__main__':
@@ -67,10 +71,6 @@ if __name__ == '__main__':
         '/': {
             'tools.sessions.on': True,
             'tools.staticdir.root': os.path.abspath(os.getcwd())
-        },
-        '/static': {
-            'tools.staticdir.on': True,
-            'tools.staticdir.dir': './public'
         }
     }
     # RUN
